requirements.txt
# Core dependencies
pandas>=2.0.0
tqdm>=4.65.0

# Scraping libraries (pilih salah satu atau install semua)

# RECOMMENDED: Simple & Free (No login required)
ntscraper>=0.4.0

# ALTERNATIVE 1: Most stable (requires dummy Twitter account)
twscrape>=0.11.0

# ALTERNATIVE 2: Good balance
tweety-ns>=1.0.0

# ALTERNATIVE 3: Old but sometimes works
# snscrape>=0.7.0  # Often unstable, not recommended

# For Selenium method (fallback)
selenium>=4.15.0
webdriver-manager>=4.0.0

# Data processing & analysis
numpy>=1.24.0
python-dateutil>=2.8.0

# For sentiment analysis (next step)
# transformers>=4.35.0
# torch>=2.1.0
# scikit-learn>=1.3.0

# For visualization (next step)
# matplotlib>=3.7.0
# seaborn>=0.12.0
# plotly>=5.17.0
# wordcloud>=1.9.0

# later add-ons
# undetected-chromedriver